构建静态库

+++ PinataEngine/src/Platform/Windows/WindowsWindow.cpp
@@ -125,14 +125,14 @@ namespace Pinata {
 		glfwSetScrollCallback(m_Window, [](GLFWwindow* window, double xoffset, double yoffset)
 			{
 				WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-				MouseScrolledEvent event(xoffset, yoffset);
+				MouseScrolledEvent event((float)xoffset, (float)yoffset);
 				data.EventCallback(event);
 			});
 
 		glfwSetCursorPosCallback(m_Window, [](GLFWwindow* window, double xpos, double ypos)
 			{
 				WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-				MouseMovedEvent event(xpos, ypos);
+				MouseMovedEvent event((float)xpos, (float)ypos);
 				data.EventCallback(event);
 			});
 	}
